Refactored DataStructureConnector class
